BVTCK-42 deleted VersionNotGivenTest, since it is ony a indirect proff of the assertion. Assertion needs to be marked as non testable
