+++ tests/src/main/resources/org/hibernate/beanvalidation/tck/tests/xmlconfiguration/versioning/validation-UnknownVersionInValidationXmlTest.xml
