--- tests/src/main/resources/org/hibernate/beanvalidation/tck/tests/xmlconfiguration/versioning/validation-VersionNotGivenTest.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<validation-config
-        xmlns="http://jboss.org/xml/ns/javax/validation/configuration"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://jboss.org/xml/ns/javax/validation/configuration validation-configuration-1.1.xsd">
-
-        <!-- Not supported as of BV 1.0 -->
-        <parameter-name-provider>org.hibernate.beanvalidation.tck.tests.xmlconfiguration.versioning.CustomParameterNameProvider</parameter-name-provider>
-</validation-config>
